Move helper to MethodOutcome
